+++ CHANGELOG.md
@@ -15,6 +15,8 @@
 ### Fixed
 - Music Map: similar artist tagging now works properly for selection with multi-value artist tags in all cases.
 - UI: avoid repaint artifacts (button hover) due to bugged SMP behaviour when opening modal popups.
+- Buttons: loved button not being refreshed on playback stop while using 'Follow now playing' setting.
+- Buttons: loved button not being refreshed when changing 'Evaluate multiple tracks' setting.
 
 ## [1.1.0] - 2025-12-12
 ### Added